+++ libs/designer/src/lib/ui/Designer.tsx
@@ -18,7 +18,7 @@ import SubgraphCardNode from './CustomNodes/SubgraphCardNode';
 import Minimap from './Minimap';
 import { ButtonEdge } from './connections/edge';
 import { HiddenEdge } from './connections/hiddenEdge';
-import { PanelRoot } from './panel/panelRoot_TEMP';
+import { PanelRoot } from './panel/panelRoot';
 import { setLayerHostSelector } from '@fluentui/react';
 import { PanelLocation } from '@microsoft/designer-ui';
 import type { WorkflowNodeType } from '@microsoft/utils-logic-apps';
